Generic template for STM32F4 project
This has a working Makefile, linker script, openocd configuration,
use of a local makefile to point to compiler and ST peripheral
library, and a working startup and system file to configure the
clocks and such.

In it's current form, this project still has some architecture
holdover from the old ST way of doing things.  I may move some
things with regard to architecture.

I plan to add a few more examples of setting up additional
peripherals.  At a minimum, this project should get the UART set
up with DMA transfer.  Other examples may include an A2D, capture
compare, timer interrupt, etc...

The ultimate goal is to have a base project to start from when
creating new projects so that the development time can be reduced.