Improvements on the lepton functions, communications, and makefile.local.
Lepton functions are now just exporting the raw data for debug.

Communications on the USART are now at 3 MBaud.

The makefile.local has been removed.  It is replaced with
makefile.local.awalker.  A symlink is made in the local directory to
makefile.local.  If anyone else wishes to use this...they can commit
a makefile.local.user and create a symlink to makefile.local on their
machine.