Added full_duplex_usart_dma.c.
This is the first in a set of reorganizing this code.  Instead of
having one large "hardware" file, I am breaking the code out into
smaller logical sections.  It will of course still be up to the
user to make sure that hardware conflicts...such as attempting to
map the same pins to multiple functions are avoided.