Added watchdog and packet timeout.
Packet timeout will reset the receive packet buffer after half
a second.  That way, if corrupt data was received and we get lost,
we can recover and hopefully catch the next valid packet instead
of being lost indefinitely.

The global watchdog needs to be verified.  It might be a good idea
to add functionality to the watchdog to run totally independent
instead of having to call the tickle function from an outside timer
function like I am doing now...